simplistic CLI test harness script
This basic script allows developer testing of device flow logins
and basic REST API calls using the device credentials. This is
suitable for use on a development sandbox. It is not a
production-ready CLI tool.

It naively saves the device token for a server hostname inside
~/.credenza/credentials.json

Example invocations:

  % credenza-device-helper login test-hostname

  % credenza-device-helper get https://test-hostname/authn/session --pretty

  % credenza-device-helper post https://test-hostname/foo --json '{"test": "input"}'

Run `credenza-device-helper help` for more usage info, or
read the source. There isn't that much to it.